feat: federation merge logic
